Add syntax module.
This module implements the object representation of the language by creating node instances from parsed token groups. The syntax is defined with the constraints:

*  delimiter == ":"
*  token_size == 1
*  rtl == true

The syntax "tree" is represented as a linked list, starting with a "root" node and zero or more "child" nodes.

*  A "root" node is created from three values: "node", "edge", and "property".
*  A "child" node is created from two values: "node" and "edge".

The three value types are as follows:

*  A "node" represents the entity/resource that is being described.
*  A "edge" represents the relationship between the "node" and the "property" when a "root" node. When a "child" node, this relationship is between the "node" and "edge".
*  A "property" represents a piece of information used to identify the "node".